fix(notes): slim paste recovery after Codex review bloat
Collapse the 40-round Lexical serialization/domain rewrite back to the
original caret-loss recovery: intercept without selection, setMarkdown
merge when there is no caret, and keep insertMarkdown when there is one.
